+++ captcha.go
@@ -1,7 +1,10 @@
 // Package captcha provides an easy to use
 package captcha
 
-import "fmt"
+import (
+	"fmt"
+	"image/color"
+)
 
 type Result struct {
 	// A random string or the result of an operation.
@@ -16,7 +19,7 @@ func CreateByText(option OptionText) (*Result, error) {
 	opt := getOptionByText(option)
 	text := opt.randomText()
 
-	data, err := createCaptcha(text, opt)
+	data, err := opt.createCaptcha(text)
 	if err != nil {
 		return nil, err
 	}
@@ -31,25 +34,9 @@ func CreateByText(option OptionText) (*Result, error) {
 // It will return a captcha with an operation like 1 + 1.
 func CreateByMath(option OptionMath) (*Result, error) {
 	opt := getOptionByMath(option)
-	min := mathMinDefault
-	if opt.MathMin != nil {
-		min = *opt.MathMin
-	}
 
-	max := mathMaxDefault
-	if opt.MathMax != nil {
-		max = *opt.MathMax
-	}
-
-	var operator matchOperator
-	if opt.MathOperator != nil {
-		operator = *opt.MathOperator
-	} else {
-		operator = randomOperation()
-	}
-
-	resultMath := generateMathOperation(&min, &max, &operator)
-	data, err := createCaptcha((*resultMath).Equation, opt)
+	resultMath := opt.generateMathOperation()
+	data, err := opt.createCaptcha(resultMath.Equation)
 	if err != nil {
 		return nil, err
 	}
@@ -60,35 +47,20 @@ func CreateByMath(option OptionMath) (*Result, error) {
 	}, nil
 }
 
-func createCaptcha(text string, opt *option) (string, error) {
-	width := widthDefault
-	if opt.Width != nil {
-		width = *opt.Width
+func (opt *option) createCaptcha(text string) (string, error) {
+	result := fmt.Sprintf("<svg xmlns=\"http://www.w3.org/2000/svg\" width=\"%v\" height=\"%v\" viewBox=\"0,0,%v,%v\" style=\"transform: rotateX(180deg)\">", opt.width, opt.height, opt.width, opt.height)
+	if opt.backgroundColor != color.Transparent {
+		r, g, b, a := opt.backgroundColor.RGBA()
+		result += fmt.Sprintf("<rect fill=\"rgba(%v, %v, %v, %v)\" width=\"100%%\" height=\"100%%\"/>", r>>8, g>>8, b>>8, a>>8)
 	}
 
-	height := heightDefault
-	if opt.Height != nil {
-		height = *opt.Height
-	}
-
-	bg := ""
-	if opt.BackgroundColor != nil {
-		isColor := true
-		opt.IsColor = &isColor
-		bg = *opt.BackgroundColor
-	}
-
-	result := fmt.Sprintf("<svg xmlns=\"http://www.w3.org/2000/svg\" width=\"%v\" height=\"%v\" viewBox=\"0,0,%v,%v\" style=\"transform: rotateX(180deg)\">", width, height, width, height)
-	if bg != "" {
-		result += fmt.Sprintf("<rect fill=\"%v\" width=\"100%%\" height=\"100%%\"/>", bg)
-	}
-
-	lineNoise := opt.drawLineNoise()
+	lineNoise := opt.drawCurve()
+	noise := opt.drawNoise()
 	pathText, err := opt.drawText(text)
 	if err != nil {
 		return "", err
 	}
-	result += fmt.Sprintf("%v%v</svg>", lineNoise, pathText)
+	result += fmt.Sprintf("%v%v%v</svg>", lineNoise, pathText, noise)
 
 	return result, nil
 }