+++ fonts/fonts.go
@@ -1,0 +1,6 @@
+package fonts
+
+import _ "embed"
+
+//go:embed Comismsh.ttf
+var Comismsh []byte