+++ CHANGELOG.md
@@ -14,3 +14,14 @@
 ### Add:
 
 - Document package
+
+## [1.1.0] - 2024-03-26
+
+### Fix:
+
+- The random color error is not working.
+
+### Feature:
+
+- Add generate noise.
+- Add the option to choose a font.